refactor(i18n): Address PR comments
