+++ lib/components/narrative/default/default-itinerary.js
@@ -7,10 +7,10 @@ import FieldTripGroupSize from '../../admin/field-trip-itinerary-group-size'
 import NarrativeItinerary from '../narrative-itinerary'
 import ItineraryBody from '../line-itin/connected-itinerary-body'
 import SimpleRealtimeAnnotation from '../simple-realtime-annotation'
+import { FormattedDuration } from '../../util/format-duration'
+import { FormattedTime } from '../../util/format-time'
 
 import ItinerarySummary from './itinerary-summary'
-import { FormattedDuration } from './format-duration'
-import { FormattedTime } from './format-time'
 
 const { calculateFares, isBicycle, isMicromobility, isTransit } = coreUtils.itinerary
