+++ lib/components/narrative/line-itin/itin-summary.js
@@ -6,8 +6,8 @@ import { connect } from 'react-redux'
 import { FormattedNumber, FormattedMessage } from 'react-intl'
 
 import { ComponentContext } from '../../../util/contexts'
-import { FormattedDuration } from '../default/format-duration'
-import { FormattedTime } from '../default/format-time'
+import { FormattedDuration } from '../../util/format-duration'
+import { FormattedTime } from '../../util/format-time'
 
 // TODO: make this a prop
 const defaultRouteColor = '#008'
@@ -145,7 +145,10 @@ export class ItinerarySummary extends Component {
               />
               <span> &bull; </span>
             </span>}
-            {Math.round(caloriesBurned)} <FormattedMessage id='common.itineraryDescriptions.calories' />
+            <FormattedMessage
+              id='common.itineraryDescriptions.calories'
+              values={{calories: Math.round(caloriesBurned)}}
+            />
           </Detail>
 
           {/* Number of transfers, if applicable */}
