+++ lib/components/narrative/narrative-itineraries-header.js
@@ -80,10 +80,12 @@ export default function NarrativeItinerariesHeader ({
             {errors.length > 0 && (
               <IssueButton onClick={onToggleShowErrors}>
                 <Icon style={{fontSize: 11, marginRight: 2}} type='warning' />
-                <span><FormattedMessage
-                  id='components.NarrativeItinerariesHeader.numIssues'
-                  values={{issueNum: errors.length}}
-                /></span>
+                <span>
+                  <FormattedMessage
+                    id='components.NarrativeItinerariesHeader.numIssues'
+                    values={{issueNum: errors.length}}
+                  />
+                </span>
               </IssueButton>
             )}
           </div>
