+++ lib/components/narrative/realtime-annotation.js
@@ -3,7 +3,7 @@ import React, { Component } from 'react'
 import { Button, OverlayTrigger, Popover } from 'react-bootstrap'
 import { FormattedList, FormattedMessage } from 'react-intl'
 
-import { FormattedDuration } from './default/format-duration'
+import { FormattedDuration } from '../util/format-duration'
 
 export default class RealtimeAnnotation extends Component {
   static propTypes = {
@@ -41,9 +41,6 @@ export default class RealtimeAnnotation extends Component {
                   normalDuration: (
                     <FormattedDuration
                       duration={realtimeEffects.normalDuration}
-                      // FIXME: move id and zeroHoursId inside component.
-                      id='common.time.tripDurationFormat'
-                      zeroHoursId='common.time.tripDurationFormatZeroHours'
                     />
                   ),
                   routes: (
