+++ lib/components/narrative/tabbed-itineraries.js
@@ -8,9 +8,8 @@ import { connect } from 'react-redux'
 import * as narrativeActions from '../../actions/narrative'
 import { ComponentContext } from '../../util/contexts'
 import { getActiveSearch, getRealtimeEffects } from '../../util/state'
-
-import { FormattedDuration } from './default/format-duration'
-import { FormattedTime } from './default/format-time'
+import { FormattedDuration } from '../util/format-duration'
+import { FormattedTime } from '../util/format-time'
 
 const { calculateFares, calculatePhysicalActivity } = coreUtils.itinerary
 
@@ -128,12 +127,14 @@ class TabButton extends Component {
         key={`tab-button-${index}`}
         onClick={this._onClick}
       >
-        <div className='title'><span>
-          <FormattedMessage
-            id='components.TabbedItineraries.optionNumber'
-            values={{optionNum: index + 1}}
-          />
-        </span></div>
+        <div className='title'>
+          <span>
+            <FormattedMessage
+              id='components.TabbedItineraries.optionNumber'
+              values={{optionNum: index + 1}}
+            />
+          </span>
+        </div>
         <div className='details'>
           {/* The itinerary duration in hrs/mins */}
           <FormattedDuration duration={itinerary.duration} />
@@ -165,7 +166,10 @@ class TabButton extends Component {
                 useMaxFare: maxTNCFare && maxTNCFare > minTNCFare ? 'true' : 'false'
               }}
             /> &bull; </span> : ''}
-            {Math.round(caloriesBurned)} <FormattedMessage id='common.itineraryDescriptions.calories' />
+            <FormattedMessage
+              id='common.itineraryDescriptions.calories'
+              values={{calories: Math.round(caloriesBurned)}}
+            />
           </span>
 
           {/* The 'X tranfers' line, if applicable */}
