+++ lib/components/util/format-time.js
